+++ .github/actions/setup-bun/action.yml
@@ -6,5 +6,3 @@ runs:
   steps:
     - name: Setup Bun
       uses: oven-sh/setup-bun@v1
-      with:
-        bun-version: '1.0.26'
