+++ lib/nns-utils/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["./**/*.ts", "./**/*.d.ts"],
+  "include": ["./**/*.ts"],
   "compilerOptions": {
     "module": "ESNext",
     "moduleResolution": "Node",
     "target": "ESNext",
-    "allowJs": true
+    "noEmit": true,
+    "types": ["@types/bun"]
   }
 }
