+++ package.json
@@ -39,12 +39,12 @@
     "@angular-eslint/template-parser": "^17.3.0",
     "@angular/cli": "^17.3.3",
     "@angular/compiler-cli": "^17.3.3",
-    "@hadronous/pic": "0.8.1",
+    "@hadronous/pic": "0.9.0-b0",
+    "@types/bun": "^1.1.1",
     "@types/node": "^20.0.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
     "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "bun-types": "^1.0.3",
     "dotenv": "^16.4.5",
     "dotenv-cli": "^7.4.1",
     "eslint": "^8.0.1",
