+++ src/backend/integration/bunfig.toml
@@ -1,0 +1,2 @@
+[test]
+preload = ["./src/preload.ts"]
