--- src/backend/integration/jest.config.js
@@ -1,12 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: 'ts-jest/presets/js-with-ts',
-  testEnvironment: 'node',
-  moduleNameMapper: {
-    '^@cg/(.*)$': '<rootDir>/../../../lib/$1',
-  },
-  globalSetup: '<rootDir>/src/global-setup.ts',
-  globalTeardown: '<rootDir>/src/global-teardown.ts',
-  testTimeout: 30_000,
-  detectOpenHandles: true,
-};
