+++ src/backend/integration/package.json
@@ -2,13 +2,13 @@
   "name": "backend-integration",
   "private": true,
   "scripts": {
-    "test": "jest"
+    "test": "tsc && bun test --timeout 60000"
   },
   "devDependencies": {
     "@cg/backend": "workspace:*",
     "@cg/nns-utils": "workspace:*",
     "@types/jest": "^29.5.12",
-    "date-fns-tz": "^2.0.1",
+    "date-fns-tz": "^3.1.3",
     "jest": "^29.5.0",
     "ts-jest": "^29.1.2"
   }
