--- src/backend/integration/src/global-setup.ts
@@ -1,12 +1,0 @@
-import { PocketIcServer } from '@hadronous/pic';
-
-module.exports = async function (): Promise<void> {
-  const pic = await PocketIcServer.start({
-    showCanisterLogs: false,
-    showRuntimeLogs: false,
-  });
-  const url = pic.getUrl();
-
-  process.env.PIC_URL = url;
-  global.__PIC__ = pic;
-};
