--- src/backend/integration/src/global-teardown.ts
@@ -1,3 +1,0 @@
-module.exports = async function () {
-  await global.__PIC__.stop();
-};
