+++ src/backend/integration/src/support/wasm.ts
@@ -2,6 +2,7 @@ import { type CanisterFixture, type PocketIc } from '@hadronous/pic';
 import { type _SERVICE, idlFactory } from '@cg/backend';
 import { resolve } from 'node:path';
 import { controllerIdentity } from './identity';
+import { Principal } from '@dfinity/principal';
 
 export const BACKEND_WASM_PATH = resolve(
   __dirname,
@@ -10,18 +11,16 @@ export const BACKEND_WASM_PATH = resolve(
   '..',
   '..',
   '..',
-  'target',
-  'wasm32-unknown-unknown',
-  'release',
-  'backend_impl.wasm',
+  '.dfx',
+  'local',
+  'canisters',
+  'backend',
+  'backend.wasm.gz',
 );
 
 export async function setupBackendCanister(
   pic: PocketIc,
-  initialDate: Date = new Date(),
 ): Promise<CanisterFixture<_SERVICE>> {
-  await pic.setTime(initialDate.getTime());
-
   const fixture = await pic.setupCanister<_SERVICE>({
     idlFactory,
     wasm: BACKEND_WASM_PATH,
@@ -32,4 +31,18 @@ export async function setupBackendCanister(
   await pic.tick(2);
 
   return fixture;
+}
+
+export async function resetBackendCanister(
+  pic: PocketIc,
+  canisterId: Principal,
+): Promise<void> {
+  await pic.reinstallCode({
+    canisterId,
+    wasm: BACKEND_WASM_PATH,
+    sender: controllerIdentity.getPrincipal(),
+  });
+
+  // make sure init timers run
+  await pic.tick(2);
 }
