+++ src/backend/integration/src/types.d.ts
@@ -1,12 +1,9 @@
-import { PocketIcServer } from '@hadronous/pic';
-
 declare global {
-  // eslint-disable-next-line no-var
-  declare var __PIC__: PocketIcServer;
-
   namespace NodeJS {
     interface ProcessEnv {
       PIC_URL: string;
     }
   }
 }
+
+export {};
